+++ src/lib/components/BaseButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BaseButton';
