+++ src/lib/components/CardButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CardButton';
