+++ src/lib/components/CountButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CountButton';
