+++ src/lib/components/MoreButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MoreButton';
