+++ src/lib/components/SliderButton/SliderButton.module.scss
