+++ src/lib/components/SliderButton/SliderButton.tsx
