+++ src/lib/components/SliderButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SliderButton';
