+++ src/lib/components/buttons/index.ts
@@ -1,8 +1,3 @@
-export * from './BaseButton';
-export * from './CountButton';
-export * from './MoreButton';
-export * from './OperationButton';
-export * from './CardButton';
 export * from './HeaderSigninButton';
 export * from './HeaderSignupButton';
 export * from './IconButton';
