+++ src/lib/stories/CountButton.stories.ts
@@ -1,4 +1,4 @@
-import { CountButton } from '@/components/buttons';
+import CountButton from '@/components/CountButton';
 
 import type { Meta, StoryObj } from '@storybook/react';
